service change: start result now returns a result
