+++ src/app/items/data-access/get-results.service.ts
@@ -6,29 +6,7 @@ import { isRouteWithSelfAttempt, FullItemRoute } from 'src/app/models/routing/it
 import { appConfig } from 'src/app/utils/config';
 import { decodeSnakeCaseZod } from 'src/app/utils/operators/decode';
 import { z } from 'zod';
-import { userBaseSchema, withGroupId } from 'src/app/groups/models/user';
-
-export interface Result {
-  attemptId: string,
-  latestActivityAt: Date,
-  startedAt: Date|null,
-  score: number,
-  validated: boolean,
-}
-
-const resultsSchema = z.array(
-  z.object({
-    id: z.string(),
-    allowsSubmissionsUntil: z.coerce.date(),
-    createdAt: z.coerce.date(),
-    endedAt: z.coerce.date().nullable(),
-    latestActivityAt: z.coerce.date(),
-    scoreComputed: z.number(),
-    startedAt: z.coerce.date().nullable(),
-    validated: z.boolean(),
-    userCreator: withGroupId(userBaseSchema).optional(),
-  })
-);
+import { Result, attemptResultSchema, resultFromFetchedResult } from '../models/attempts';
 
 @Injectable({
   providedIn: 'root'
@@ -43,14 +21,8 @@ export class GetResultsService {
         params: isRouteWithSelfAttempt(item) ? { attempt_id: item.attemptId } : { parent_attempt_id: item.parentAttemptId },
       })
       .pipe(
-        decodeSnakeCaseZod(resultsSchema),
-        map(results => results.map(r => ({
-          attemptId: r.id,
-          latestActivityAt: r.latestActivityAt,
-          startedAt: r.startedAt,
-          score: r.scoreComputed,
-          validated: r.validated,
-        }))),
+        decodeSnakeCaseZod(z.array(attemptResultSchema)),
+        map(results => results.map(resultFromFetchedResult)),
       );
   }
