+++ src/app/items/models/attempts.ts
@@ -1,13 +1,42 @@
+import { userBaseSchema, withGroupId } from 'src/app/groups/models/user';
 import { canCurrentUserViewContent, ItemPermWithView } from './item-view-permission';
+import { z } from 'zod';
+
+export const attemptResultSchema = z.object({
+  id: z.string(),
+  allowsSubmissionsUntil: z.coerce.date(),
+  createdAt: z.coerce.date(),
+  endedAt: z.coerce.date().nullable(),
+  latestActivityAt: z.coerce.date(),
+  scoreComputed: z.number(),
+  startedAt: z.coerce.date().nullable(),
+  validated: z.boolean(),
+  userCreator: withGroupId(userBaseSchema).optional(),
+});
+
+export interface Result {
+  attemptId: string,
+  latestActivityAt: Date,
+  startedAt: Date|null,
+  score: number,
+  validated: boolean,
+  allowsSubmissionsUntil: Date,
+}
+
+export function resultFromFetchedResult(result: z.infer<typeof attemptResultSchema>): Result {
+  return {
+    attemptId: result.id,
+    latestActivityAt: result.latestActivityAt,
+    startedAt: result.startedAt,
+    score: result.scoreComputed,
+    validated: result.validated,
+    allowsSubmissionsUntil: result.allowsSubmissionsUntil
+  };
+}
 
 export const defaultAttemptId = '0'; // to be used as parent attempt for root items
 
-interface Result {
-  latestActivityAt: Date,
-  startedAt: Date|null,
-}
-
-export function bestAttemptFromResults<T extends Result>(results: T[]): T|null {
+export function bestAttemptFromResults<T extends { latestActivityAt: Date, startedAt: Date|null }>(results: T[]): T|null {
   if (results.length === 0) return null;
 
   // pick the one which is started and with the greatest latest_activity_at
