+++ src/app/items/models/item-data.ts
@@ -1,7 +1,7 @@
 import { FullItemRoute } from 'src/app/models/routing/item-route';
 import { BreadcrumbItem } from '../data-access/get-breadcrumb.service';
 import { Item } from '../../data-access/get-item-by-id.service';
-import { Result } from '../data-access/get-results.service';
+import { Result } from '../models/attempts';
 
 /**
  * All fetched data related to one item
