+++ src/app/items/services/result-fetching.service.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
 import { Item } from 'src/app/data-access/get-item-by-id.service';
 import { FullItemRoute, isRouteWithSelfAttempt } from 'src/app/models/routing/item-route';
-import { GetResultsService, Result } from '../data-access/get-results.service';
+import { GetResultsService } from '../data-access/get-results.service';
 import { EMPTY, Observable, map, of, switchMap } from 'rxjs';
 import { canCurrentUserViewContent } from '../models/item-view-permission';
-import { bestAttemptFromResults, implicitResultStart } from '../models/attempts';
+import { Result, bestAttemptFromResults, implicitResultStart } from '../models/attempts';
 import { ResultActionsService } from 'src/app/data-access/result-actions.service';
 
 export function canFetchResults(item: Item): boolean {
@@ -39,10 +39,7 @@ export class ResultFetchingService {
         const attemptId = isRouteWithSelfAttempt(itemRoute) ? itemRoute.attemptId : itemRoute.parentAttemptId;
         if (!attemptId) return EMPTY; // unexpected
         return this.resultActionsService.start(itemRoute.path.concat([ itemRoute.id ]), attemptId).pipe(
-          map(() => {
-            const result = { attemptId, latestActivityAt: new Date(), startedAt: new Date(), score: 0, validated: false };
-            return { results: [ ...results, result ], currentResult: result };
-          })
+          map(result => ({ results: [ ...results, result ], currentResult: result })),
         );
       }),
     );
